Add lsp mode for kotlin, get it semi working
The daemon is up, at the least. Getting errors with null hashmaps

Odd... sometimes it works, sometimes it times out, sometimes it
reports a ton of errors. Still looks very promising!

One thing that would be nice is if it could jump us to vendor
libraries. Jumping to Kotlin compiler internals would be dope