Add x250.org
Since we're actually using emacs as a first class code editor at work
it makes sense to start pulling out things into separate profiles.

Unfortunately there's a one time first step: selecting the right
profile to build on "boot": This has to be done every time we nuke the
profile, but that's acceptable.

To get it done automatically we need to switch on something, but
that's not really that important, it's pretty easy for me to tell what
context I'm on.