Add go
Now I'm starting to think I miss the model of a single file that gets
tangled.

All I want to do is infer the machine I'm on and use that to load in
config. The "require" setup doesn't do that, but I could see
use-package declarations with a default of "no config" (for new
machines).

I need a way to infer or set what machine I'm on. Then anytime I want
to switch, I do so.