Add nonguix to load path
Not really sure right now how to do this the right way: this runs the
risk of nonguix being garbage collected (or worse, inaccurate to the
running system).

I tried for quite a long time to get guix repl to behave as a repl for
geiser, but ran into a few problems.