Don't use use-rerender hook
There's a big problem with the use-rerender hook: if the component
passed in is not the rendering component (if there's an orchestrator
component), then only the orchestrator component will be re-rendered,
not any of the children (unless they do something bad like
":should-component-update 'always")

Also, drop the old use-process. Buffers are the way to go, despite the
shortcomings we've seen here.