Use subscribe-to-attribute for use-package-conditionally
use-package-conditionally is a special case of subscribe-to-attribute:
it's really common for me to register a package based on the
environment. It might be useful to turn it on and off.

Basically these are all kludges until elpaca has lockfiles. Until then
if we want org (which takes _forever_ to clone), we can just flip has-org